# Conflicts: #	core/src/com/nuclearthinking/myheroagency/controller/ScreenEnum.java #	core/src/com/nuclearthinking/myheroagency/view/AbstractScreen.java #	core/src/com/nuclearthinking/myheroagency/view/HomeScreen.java

+++ core/src/com/nuclearthinking/myheroagency/model/Player.java
@@ -12,7 +12,7 @@ class Player implements Serializable {
         return level;
     }
 
-    public void setLevel(int level) {
+    public void setLevel(byte level) {
         this.level = level;
     }
